+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- The Discovery feature is not functioning correctly because the release package is missing the `composer.json` file.
+
 ## [1.0.0] - 2025-12-27
 
 ### Added